chore(suite-native): move passphrase discovery selectors
- aiming to reduce footprint of device-authorization package on passphrase package. Shall be removed completely, but that means refactoring more of the navigation logic so it will be done as a follow up
