+++ suite-native/device-authorization/src/deviceAuthorizationSlice.ts
@@ -1,10 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import {
-    DeviceRootState,
-    DiscoveryRootState,
-    selectDiscoveryByDevicePath,
-} from '@suite-common/wallet-core';
 import { UI } from '@trezor/connect';
 
 import {
@@ -101,86 +96,6 @@ export const selectInputPassphraseOnDevice = (state: DeviceAuthorizationRootStat
 export const selectCheckPassphraseOnDevice = (state: DeviceAuthorizationRootState) =>
     state.deviceAuthorization.deviceAuthorizationStep ===
     DeviceAuthorizationStep.CheckPassphraseOnDevice;
-
-export const selectHasPassphraseError = (state: DiscoveryRootState & DeviceRootState) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    return (
-        discovery?.isAddingExistingWallet &&
-        ['failed', 'cancelled', 'passphrase-mismatch'].includes(discovery.status)
-    );
-};
-
-export const selectHasVerificationCancelledError = (
-    state: DiscoveryRootState & DeviceRootState,
-) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    return discovery?.status === 'cancelled';
-};
-
-export const selectHasPassphraseMismatchError = (state: DiscoveryRootState & DeviceRootState) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    return discovery?.status === 'passphrase-mismatch';
-};
-
-export const selectIsCreatingNewPassphraseWallet = (
-    state: DiscoveryRootState & DeviceRootState,
-) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    return discovery?.isAddingHiddenWallet;
-};
-
-export const isPassphraseDeviceLoadingDone = (
-    state: DiscoveryRootState & DeviceRootState & DeviceAuthorizationRootState,
-) => {
-    if (!state.device.selectedDevice?.state) {
-        return false;
-    }
-
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    if (!discovery || !discovery.isAddingHiddenWallet) {
-        return false;
-    }
-
-    return (
-        state.deviceAuthorization.deviceAuthorizationStep !==
-        DeviceAuthorizationStep.PassphraseRequested
-    );
-};
-
-export const selectPassphraseDeviceNotEmpty = (state: DiscoveryRootState & DeviceRootState) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    if (!discovery || !discovery.isAddingHiddenWallet) {
-        return null;
-    }
-
-    switch (discovery.status) {
-        case 'confirm-empty-passphrase':
-            return false;
-        case 'complete':
-            return true;
-        default:
-            return null;
-    }
-};
-
-export const selectPassphraseDiscoveryCompleted = (state: DiscoveryRootState & DeviceRootState) => {
-    const discovery = selectDiscoveryByDevicePath(state, state.device.selectedDevice?.path);
-
-    if (!discovery || !discovery.isAddingHiddenWallet) {
-        return null;
-    }
-
-    return (
-        discovery.status === 'complete' ||
-        (discovery.status === 'progress' && discovery.hasLoadedAnyNonEmptyAccount)
-    );
-};
 
 export const { checkPassphraseOnDevice } = deviceAuthorizationSlice.actions;
